ada: VAST: treewalker improvements
Implement two basic checks: Check that N_Error nodes cannot appear in the
tree (because VAST is not called when the source is illegal).
Check that every node has a parent, except for certain nodes where
we check the opposite. (We do not yet check that the parent pointers
actually point to the right node.)

Minor improvements. Flags for controlling debugging outputs and the like.

Capability to enable/disable individual checks. The intent is to
implement that only when needed -- i.e. when VAST finds a bug, and we
have not yet fixed the bug.

gcc/ada/ChangeLog:

	* vast.adb: Implement two checks. Improve debugging
	outputs.