ada: Factorize some duplicate code
Process_Subtype can be passed either a subtype indication or a subtype
mark. Before this patch, it branched directly depending on the kind of
the argument, but there actually was processing common to the two
branches like resolving the subtype mark. This patch factorizes this
common processing out of the if statement.

gcc/ada/ChangeLog:

	* sem_ch3.adb (Process_Subtype): Factorize code.