ada: Small cleanup in instantiation of generic bodies
This removes the code dealing with generic main units from the body of the
instantiation routine, namely Instantiate_Bodies, and replaces it by tests
done earlier in the processing.

The test added to Need_Subprogram_Instance_Body is already present in the
twin predicate Needs_Body_Instantiated.

gcc/ada/ChangeLog:

	* inline.adb (Instantiate_Body): Do not call Add_Scope_To_Clean if
	the main unit is generic.
	(Instantiate_Bodies): Do not deal with generic main units here.
	* sem_ch12.adb (Need_Subprogram_Instance_Body): Return false if the
	main unit is generic.