testsuite: revert Fortran change
Revert part of my change from r15-8973-g1307de1b4e7d5e; as Harald points
out, the comment explains why this is there. It's a hack but it needs to
stay for now. (I did have this marked as a TODO in my branch and didn't
leave a proper note as to why, so it's my fault.)

gcc/testsuite/ChangeLog:

	* gfortran.dg/cray_pointers_2.f90: Restore whitespace.